refactor: move the command list
